+++ CHANGELOG.md
@@ -4,3 +4,13 @@
 ### Added
 - First commit;
 - Quanlse SDK contains toolkits, tutorials, examples.
+
+## [1.1.0] - April-19-2021
+### Added
+- New 1-qubit optimizer using GRAPE (GRadient Ascent Pulse Engineering) Algorithm;
+
+- Noisy 1-qubit simulator including readout simulation;
+
+- Trapped ion: 1-qubit optimizer and the  two-qubit MS gate optimizer in trapped ion 
+
+  quantum computing.